browser(clock, perf): review round on #86: lock-free p2d ring, fire outside the locks, condvar fallback, jitter test, one clock
Eight findings from the Opus review, all in:
- perf.rs: a fixed ring of 1024 atomics with an atomic head; a render
  does two atomic stores, no lock, no allocation. Summary copies at most
  the ring.
- clock.rs: asked= fired= as a pair; deadlines collected under the locks
  and fired with both released, so a waker may push again; Windows falls
  back to the condvar wait when a kernel object is refused, and a refused
  SetWaitableTimer waits on the event with a bounded timeout, never a
  sleep loop; a jitter test (fifty deadlines 4 ms apart, p95 error under
  2 ms) and a re-entrancy test.
- pump.rs: its condvar clock goes; schedule_pump waits on clock.rs, and
  timer_fired= reads the clock's count. One thread named
  surya-browser-clock.
- Cargo.toml: Win32_Security, which CreateEventW and CreateWaitableTimerExW
  are gated behind; found by winprobe/, a two-file crate that checks the
  Windows modules on x86_64-pc-windows-msvc from Linux, kept in tree.

cargo test --lib: 54 passed, 0 failed. winprobe check on the MSVC target:
Finished, 0 errors.

+++ app/crates/browser/src/selftest/mod.rs
@@ -67,7 +67,7 @@ fn wheel(dy: i32) -> bool {
 struct Counts {
     cef_frames: u64,
     app_frames: u64,
-    p2d_mark: usize,
+    p2d_mark: u64,
 }
 
 fn counts() -> Counts {
